made some adjustment and team deletion is done
