--- frontend/src/css/Instructor.css
@@ -1,144 +1,0 @@
-.instructor form {
-  width: 1200px;
-  margin: 20px auto;
-}
-
-.header {
-  width: 1200px;
-  margin: 0 auto;
-}
-
-.instructor .top-display {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-
-  margin: 20px 20px;
-}
-
-.instructor h2 {
-  color: #cc813f;
-  font-weight: 800;
-  margin-right: 20px;
-}
-
-.instructor h3 {
-  color: #ffffff;
-  font-weight: 800;
-  margin: 20px;
-}
-
-.instructor .top-display .team-name {
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-}
-
-.instructor .top-display .team-name {
-  background-color: #464545;
-  height: 35px;
-  width: 800px;
-}
-
-.instructor .top-display .team-name input {
-  border: none;
-  border-bottom: solid #a8a8a8 1px;
-  background-color: #464545;
-
-  height: 32px;
-  width: 750px;
-
-  color: white;
-  font-size: 12pt;
-}
-
-.instructor ul {
-  background-color: #464545;
-  height: auto;
-  overflow: auto;
-  margin-bottom: 20px;
-  padding-bottom: 20px;
-}
-
-.instructor ul li {
-  margin: 0px 50px;
-  color: #ffffff;
-}
-/* 
-.instructor li label {
-  color: #ffffff;
-  font-weight: 200;
-  font-size: 12pt;
-}
-
-.instructor li input {
-  margin-right: 15px;
-  margin-left: 10px;
-} */
-
-.instructor .buttons {
-  display: flex;
-  flex-direction: row;
-  justify-content: flex-end;
-
-  gap: 20px;
-}
-
-.buttons > * {
-  color: #cc813f;
-  background-color: #464545;
-
-  width: 100px;
-  height: 30px;
-
-  cursor: pointer;
-  border: none;
-}
-
-.instructor .buttons button:hover {
-  background-color: #cc813f;
-  color: #464545;
-}
-
-.instructor .buttons button:active {
-  opacity: 0.7;
-}
-
-/* detele and edit */
-.instructor .delEdit {
-  display: flex;
-  flex-direction: row;
-  justify-content: flex-end;
-
-  gap: 20px;
-}
-
-.delEdit > .delete,
-.delEdit > .edit,
-.delEdit > .more {
-  width: 100px;
-  height: 30px;
-
-  cursor: pointer;
-  border: none;
-}
-
-.instructor .delEdit button:active {
-  opacity: 0.7;
-}
-
-.delete {
-  background-color: red;
-  color: #ffffff;
-}
-
-.edit {
-  background-color: gray;
-  color: #ffffff;
-}
-
-.more {
-  background-color: gray;
-  color: #ffffff;
-}
-
